refactor(xtts): use position embedding from tortoise
